OS-62: Improved code
