+++ OpenImis.ModulesV2/FeedbackModule/Repositories/IFeedbackRepository.cs
@@ -8,7 +8,6 @@ namespace OpenImis.ModulesV2.FeedbackModule.Repositories
     public interface IFeedbackRepository
     {
         List<FeedbackResponseModel> Get(string officerCode);
-        string GetLoginNameByUserUUID(Guid userUUID);
         int Post(FeedbackRequest feedbackClaim);
     }
 }
