+++ OpenImis.ModulesV2/Helpers/Extensions.cs
@@ -33,5 +33,19 @@ namespace OpenImis.ModulesV2.Helpers
                 throw new Exception("An error occurred", ex);
             }
         }
+
+        public static decimal? ToNullableDecimal(this string s)
+        {
+            decimal d;
+            if (decimal.TryParse(s, out d)) return d;
+            return null;
+        }
+
+        public static float? ToNullableFloat(this string s)
+        {
+            float f;
+            if (float.TryParse(s, out f)) return f;
+            return null;
+        }
     }
 }
