+++ OpenImis.ModulesV2/Helpers/OutputParameter.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
+﻿using System.Data;
 using System.Data.SqlClient;
-using System.Text;
 
-namespace OpenImis.ModulesV2.PolicyModule.Helpers
+namespace OpenImis.ModulesV2.Helpers
 {
     public static class OutputParameter
     {
