+++ OpenImis.ModulesV2/PolicyModule/Logic/IPolicyRenewalLogic.cs
@@ -10,6 +10,5 @@ namespace OpenImis.ModulesV2.PolicyModule.Logic
         List<GetPolicyRenewalModel> Get(string officerCode);
         int Post(PolicyRenewalModel policy);
         int Delete(Guid uuid);
-        string GetLoginNameByUserUUID(Guid userUUID);
     }
 }
