+++ OpenImis.ModulesV2/PolicyModule/Repositories/IPolicyRenewalRepository.cs
@@ -9,6 +9,5 @@ namespace OpenImis.ModulesV2.PolicyModule.Repositories
         List<GetPolicyRenewalModel> Get(string officerCode);
         int Post(PolicyRenewalModel policy);
         int Delete(Guid uuid);
-        string GetLoginNameByUserUUID(Guid userUUID);
     }
 }
