+++ OpenImis.ModulesV2/ReportModule/Logic/IReportLogic.cs
@@ -8,6 +8,5 @@ namespace OpenImis.ModulesV2.ReportModule.Logic
         FeedbackReportModel GetFeedbackStats(ReportRequestModel feedbackRequestModel, string officerCode);
         RenewalReportModel GetRenewalStats(ReportRequestModel renewalRequestModel, string officerCode);
         EnrolmentReportModel GetEnrolmentStats(ReportRequestModel enrolmentRequestModel, string officerCode);
-        string GetLoginNameByUserUUID(Guid userUUID);
     }
 }
