+++ OpenImis.ModulesV2/ReportModule/Repositories/IReportRepository.cs
@@ -8,6 +8,5 @@ namespace OpenImis.ModulesV2.ReportModule.Repositories
         FeedbackReportModel GetFeedbackStats(ReportRequestModel feedbackRequestModel, string officerCode);
         RenewalReportModel GetRenewalStats(ReportRequestModel renewalRequestModel, string officerCode);
         EnrolmentReportModel GetEnrolmentStats(ReportRequestModel enrolmentRequestModel, string officerCode);
-        string GetLoginNameByUserUUID(Guid userUUID);
     }
 }
