+++ OpenImis.RestApi/Controllers/V2/FeedbackController.cs
@@ -1,13 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Diagnostics;
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using OpenImis.ModulesV2;
 using OpenImis.ModulesV2.FeedbackModule.Models;
+using OpenImis.ModulesV2.Utils;
 using OpenImis.RestApi.Security;
 
 namespace OpenImis.RestApi.Controllers.V2
@@ -53,7 +52,8 @@ namespace OpenImis.RestApi.Controllers.V2
             {
                 Guid userUUID = Guid.Parse(HttpContext.User.Claims.Where(w => w.Type == "UserUUID").Select(x => x.Value).FirstOrDefault());
 
-                string officerCode = _imisModules.GetFeedbackModule().GetFeedbackLogic().GetLoginNameByUserUUID(userUUID);
+                Repository rep = new Repository();
+                string officerCode = rep.GetLoginNameByUserUUID(userUUID);
 
                 response = _imisModules.GetFeedbackModule().GetFeedbackLogic().Get(officerCode);
             }
